Refactor ActiveWindow to separate file
